# inputs/openapi.yaml
openapi: 3.0.3
info:
  title: VMware vRA REST API Server
  version: 0.1.0
  description: |
    REST API server for VMware vRealize Automation (vRA) operations.
    
    This server provides programmatic access to VMware vRA functionality including:
    - Authentication and token management
    - Catalog item browsing and requesting
    - Deployment management and monitoring
    
    The API follows standard REST conventions and returns JSON responses.
  contact:
    name: VMware vRA CLI
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /deployments:
    get:
      tags:
        - deployments
      summary: List deployments
      description: |
        Retrieve a list of deployments with optional filtering by project and status.
        Results can be paginated for better performance.
      parameters:
        - name: project_id
          in: query
          description: Filter by project ID
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by deployment status
          required: false
          schema:
            type: string
        - name: page_size
          in: query
          description: Number of items per page (1-2000)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 100
        - name: first_page_only
          in: query
          description: Fetch only the first page
          required: false
          schema:
            type: boolean
            default: false
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentsResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/{deployment_id}:
    get:
      tags:
        - deployments
      summary: Get deployment details
      description: Retrieve detailed information about a specific deployment
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: string
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - deployments
      summary: Delete deployment
      description: |
        Delete a deployment. This operation requires explicit confirmation
        via the confirm parameter to prevent accidental deletions.
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: string
        - name: confirm
          in: query
          description: Confirmation required for deletion
          required: true
          schema:
            type: boolean
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Deployment deletion initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Confirmation required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/{deployment_id}/resources:
    get:
      tags:
        - deployments
      summary: Get deployment resources
      description: Retrieve the resources associated with a specific deployment
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: string
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Deployment resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResourcesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
          description: Indicates if the operation was successful
        message:
          type: string
          nullable: true
          description: Human-readable message about the operation
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp of the response
      required:
        - success
        - timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: false
            error_code:
              type: string
              nullable: true
              description: Machine-readable error code
            error_details:
              type: object
              nullable: true
              description: Additional error details
          required:
            - success
...
    DeploymentsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            deployments:
              type: array
              items:
                type: object
                description: Deployment data
              description: List of deployments
            total_count:
              type: integer
              description: Total number of deployments returned
            page_info:
              type: object
              nullable: true
              properties:
                page_size:
                  type: integer
                first_page_only:
                  type: boolean
                project_filter:
                  type: string
                  nullable: true
                status_filter:
                  type: string
                  nullable: true
              description: Pagination and filtering information
          required:
            - deployments
            - total_count
    DeploymentResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            deployment:
              type: object
              description: Detailed deployment information
          required:
            - deployment
    DeploymentResourcesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            resources:
              type: array
              items:
                type: object
                description: Resource data
              description: List of deployment resources
          required:
            - resources
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        VMware vRA access token obtained via the /auth/login endpoint.
        The token is automatically managed by the server after authentication.
security:
  - BearerAuth: []
...